lib: move TreeValue-specific Merge impls to backend module
Move the `MergedTreeValue`/`MergedTreeVal` aliases and the
`Merge<Option<T>>` methods that match on `TreeValue` variants to the
module where `TreeValue` is defined. This makes the merge module more
generic. It also makes the dependency between the two modules
one-directional: the backend module already depended on the merge
module for `Commit::root_tree`.
