+++ cli/src/commands/debug/object.rs
@@ -19,10 +19,10 @@ use clap::Subcommand;
 use futures::AsyncReadExt as _;
 use jj_lib::backend::CommitId;
 use jj_lib::backend::FileId;
+use jj_lib::backend::MergedTreeValue;
 use jj_lib::backend::SymlinkId;
 use jj_lib::backend::TreeId;
 use jj_lib::backend::TreeValue;
-use jj_lib::merge::MergedTreeValue;
 use jj_lib::op_store::OperationId;
 use jj_lib::op_store::ViewId;
 use jj_lib::repo_path::RepoPath;
