+++ lib/src/conflicts.rs
@@ -37,6 +37,7 @@ use crate::backend::BackendResult;
 use crate::backend::CommitId;
 use crate::backend::CopyId;
 use crate::backend::FileId;
+use crate::backend::MergedTreeValue;
 use crate::backend::SymlinkId;
 use crate::backend::TreeId;
 use crate::backend::TreeValue;
@@ -50,7 +51,6 @@ use crate::files;
 use crate::files::MergeResult;
 use crate::merge::Diff;
 use crate::merge::Merge;
-use crate::merge::MergedTreeValue;
 use crate::merge::SameChange;
 use crate::repo_path::RepoPath;
 use crate::store::Store;
