+++ lib/src/copies.rs
@@ -38,11 +38,11 @@ use crate::backend::BackendResult;
 use crate::backend::CopyHistory;
 use crate::backend::CopyId;
 use crate::backend::CopyRecord;
+use crate::backend::MergedTreeValue;
 use crate::backend::TreeValue;
 use crate::dag_walk;
 use crate::merge::Diff;
 use crate::merge::Merge;
-use crate::merge::MergedTreeValue;
 use crate::merge::SameChange;
 use crate::merged_tree::MergedTree;
 use crate::merged_tree::TreeDiffEntry;
