+++ lib/src/local_working_copy.rs
@@ -64,6 +64,7 @@ use tracing::trace_span;
 use crate::backend::BackendError;
 use crate::backend::CopyId;
 use crate::backend::FileId;
+use crate::backend::MergedTreeValue;
 use crate::backend::MillisSinceEpoch;
 use crate::backend::SymlinkId;
 use crate::backend::TreeId;
@@ -102,7 +103,6 @@ use crate::matchers::PrefixMatcher;
 use crate::matchers::UnionMatcher;
 use crate::merge::Merge;
 use crate::merge::MergeBuilder;
-use crate::merge::MergedTreeValue;
 use crate::merge::SameChange;
 use crate::merged_tree::MergedTree;
 use crate::merged_tree::TreeDiffEntry;
