+++ lib/src/tree_merge.rs
@@ -32,14 +32,14 @@ use itertools::Itertools as _;
 use crate::backend;
 use crate::backend::BackendError;
 use crate::backend::BackendResult;
+use crate::backend::MergedTreeVal;
+use crate::backend::MergedTreeValue;
 use crate::backend::TreeId;
 use crate::backend::TreeValue;
 use crate::config::ConfigGetError;
 use crate::files;
 use crate::files::FileMergeHunkLevel;
 use crate::merge::Merge;
-use crate::merge::MergedTreeVal;
-use crate::merge::MergedTreeValue;
 use crate::merge::SameChange;
 use crate::merged_tree::all_merged_tree_entries;
 use crate::object_id::ObjectId as _;
