+++ lib/tests/test_merged_tree.rs
@@ -21,6 +21,7 @@ use jj_lib::backend::CommitId;
 use jj_lib::backend::CopyHistory;
 use jj_lib::backend::CopyRecord;
 use jj_lib::backend::FileId;
+use jj_lib::backend::MergedTreeValue;
 use jj_lib::backend::TreeValue;
 use jj_lib::conflict_labels::ConflictLabels;
 use jj_lib::copies::CopiesTreeDiffEntryPath;
@@ -36,7 +37,6 @@ use jj_lib::matchers::Matcher;
 use jj_lib::matchers::PrefixMatcher;
 use jj_lib::merge::Diff;
 use jj_lib::merge::Merge;
-use jj_lib::merge::MergedTreeValue;
 use jj_lib::merged_tree::MergedTree;
 use jj_lib::merged_tree::TreeDiffEntry;
 use jj_lib::merged_tree::TreeDiffIterator;
